Using FrameRateSelection for selecting of framerate
Redesigned Selection. Moved selection menu elements from Theme to Selection.
Created SampleRateSelection and FrameRateSelection classes derived
from Selection

+++ cinelerra/recordconfig.h
@@ -85,7 +85,7 @@ public:
 // Dimensions of captured frame
 	int w, h;
 // Frame rate of captured frames
-	float in_framerate;
+	double in_framerate;
 };
